inline DefaultAttributeMap into AbstractChannel
